+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix: breaking clap change, same syntax as before for `--cities`.
 - feat: Add Mouse control for Map/Coverage lat/long position.
 - feat: Add Mouse control for tab selection.
+- refactor: Add LAT_LONG_DIFF const
 
 ### apps/1090